+++ src/main/java/roomescape/handler/GlobalExceptionHandler.java
@@ -1,4 +1,4 @@
-package roomescape.reservation.handler;
+package roomescape.handler;
 
 import java.time.format.DateTimeParseException;
 import org.slf4j.Logger;
@@ -8,7 +8,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.http.converter.HttpMessageNotReadableException;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
-import roomescape.reservation.handler.exception.CustomException;
+import roomescape.handler.exception.CustomException;
 
 @ControllerAdvice
 public class GlobalExceptionHandler {
